creacion de controladores de timetables y tests

+++ app/Http/Controllers/Timetables/DeleteTimetableController.php
@@ -5,14 +5,27 @@ namespace App\Http\Controllers\Timetables;
 use App\Http\Controllers\Controller;
 use App\Models\Timetable;
 use App\Traits\ApiFeedbackSender;
+use Illuminate\Http\Request;
 
 class DeleteTimetableController extends Controller
 {
     use ApiFeedbackSender;
 
-    public function __invoke(Timetable $timetable)
+    public function __construct()
     {
+        $this->middleware('auth:sanctum');
+    }
+
+    public function __invoke(Request $request, $id)
+    {
+        $timetable = Timetable::where('user_id', $request->user()->id)
+            ->find($id);
+        
+        if (!$timetable) {
+            return $this->sendError('Horario no encontrado', ['No se encontró el recurso solicitado'], 404);
+        }
+
         $timetable->delete();
-        return $this->sendSuccess(null, 'Horario eliminado exitosamente');
+        return $this->sendSuccess('Horario eliminado exitosamente', null);
     }
 }
